requirements/requirements.txt
click==8.1.3
requests==2.28.1
rich==13.0.0
Mastodon.py==1.8.0
dropbox==11.36.0
beautifulsoup4==4.11.1
openai==0.26